fix: line thickness is int parameter

+++ cameras/config.py
@@ -70,7 +70,7 @@ class Colors(Enum):
 class Graphics(Enum):
     ''' Class representing graphics configuration '''
     LINE_COLOR: Tuple[int] = Colors.BLACK.value
-    LINE_THICKNESS: float = float(config.get('GRAPHICS', 'LINE_THICKNESS'))
+    LINE_THICKNESS: int = int(config.get('GRAPHICS', 'LINE_THICKNESS'))
     LINE_TEXT: str = config.get('GRAPHICS', 'LINE_TEXT')
 
     TITLE: str = config.get('GRAPHICS', 'TITLE')
